requirements.txt
opencv-python==4.5.5.64
PyQt5==5.15.6
scipy==1.8.1
numpy==1.22.4
Pillow==9.1.1
scikit-image==0.19.2
matplotlib==3.5.2